feat: refactor buyer and seller form components, add new SVG assets, and implement zip code input functionality
