+++ app/buyer-form/page.tsx
@@ -1,14 +1,14 @@
 import React from "react";
-import BuyerFormBody from "./components/formComponents/BuyerFormBody";
+import SellerFromBody from "./components/SellerFormBody";
 import FormHeader from "../../components/FormComponents/FormHeader";
 
-function BuyerForm() {
+const page = () => {
   return (
     <div className="max-w-[50rem] mx-auto flex flex-col gap-8 p-6 mb-[2rem] bg-blue rounded-md text-white shadow-md">
-      <FormHeader pageTitle="What type of house is your property?" />
-      <BuyerFormBody />
+      <FormHeader pageTitle="Let's Find Your Dream Home" />
+      <SellerFromBody />
     </div>
   );
-}
+};
 
-export default BuyerForm;
+export default page;
