+++ app/seller-form/components/assets/DetachedHouseIcon.tsx
