+++ app/seller-form/components/assets/NewlyRenovatedIcon.tsx
