+++ app/seller-form/components/assets/SemiDetachedHouse.tsx
