+++ app/seller-form/components/assets/TerraceHouseCorner.tsx
