+++ app/seller-form/components/assets/TerraceHouseMiddle.tsx
