+++ app/seller-form/components/assets/WellMaintainedIcon.tsx
