+++ app/seller-form/components/formComponents/BuyerConsentForm.tsx
