+++ app/seller-form/components/formComponents/HouseCondition.tsx
