+++ app/seller-form/page.tsx
@@ -1,14 +1,14 @@
 import React from "react";
-import SellerFromBody from "./components/SellerFormBody";
+import BuyerFormBody from "./components/formComponents/BuyerFormBody";
 import FormHeader from "../../components/FormComponents/FormHeader";
 
-const page = () => {
+function BuyerForm() {
   return (
     <div className="max-w-[50rem] mx-auto flex flex-col gap-8 p-6 mb-[2rem] bg-blue rounded-md text-white shadow-md">
-      <FormHeader pageTitle="Let's Find Your Dream Home" />
-      <SellerFromBody />
+      <FormHeader pageTitle="What type of house is your property?" />
+      <BuyerFormBody />
     </div>
   );
-};
+}
 
-export default page;
+export default BuyerForm;
